refactor: refactor the modules folder
